Smart-cusine-Recommandation-main/requirements.txt
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
gunicorn==26.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
nltk==3.9.4
numpy==2.4.4
packaging==26.2
pandas==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.2.1
regex==2026.4.4
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.6
